oidc: remove restriction for callbackURLs
The original limitation required the OIDC callbackURL to start with
the Dex issuer, which prevented the use of relative issuer strings.
Relative issuer is useful if we don't know all the URLs how we can be
called.

After this change, it is the discretion of the administrator to provide
a URL that can be routed back to Dex.